+++ src/routes/habitat/components/Chat/index.js
@@ -8,18 +8,19 @@ import style from './style.scss';
 
 const ChatComponent = lazy(() => import('Components/Chat'));
 
-const Chat = ({ height, width }) => (
+const Chat = ({ height, width, habitatId }) => (
   <div
     className={style.chat}
     style={{
       height,
       maxHeight: height,
+      maxWidth: width,
       width,
     }}
   >
     <ScheduleCarousel />
     <Suspense>
-      <ChatComponent />
+      <ChatComponent channelId={habitatId} />
     </Suspense>
   </div>
 );